+++ src/Workflowish/Subcomponents/OmnibarWrapper/index.tsx
@@ -1,71 +1,44 @@
 import * as React from 'react';
-import { ItemTreeNode, TodoItemsGetSetterWithKeyedNodes } from '../../mvc/model';
+import { TransformedDataAndSetter } from '../../mvc/model';
 import { ModelContext } from '~Workflowish/mvc/context';
 import { FocusActions } from '~Workflowish/Item';
 import "./index.css"
-import { NO_MATCH, resetSearchState, searchTransformFromOmnibarState } from './Search';
 import { OmniBarState } from './States';
+import { getSpecializedProps } from './Specializations';
 
 
-
-export default (props: {
+const OmniBarWrapper = (props: {
     children: React.ReactElement,
     itemRefsDictionary: Record<string, FocusActions>,
-    getSetTodoItems: TodoItemsGetSetterWithKeyedNodes,
-    lastFocusedItem: string
+    transformedDataAndSetter: TransformedDataAndSetter,
+    lastFocusedItem: string,
 }) => {
-    const [omniBarState, setOmniBarState] = React.useState<OmniBarState>({
-        barContents: "",
-        searchState: resetSearchState()
-    });
-    const upperRootNode = React.useContext<ItemTreeNode>(ModelContext);
-    const { rootNode,
-        nMatches,
-        currentMatchId,
-        currentMatchParentChain
-    } = searchTransformFromOmnibarState(upperRootNode, omniBarState, setOmniBarState);
-    const focusOnCurrentItem = () => {
-        setOmniBarState({ barContents: "", searchState: resetSearchState() });
-        props.getSetTodoItems((oldRoot, oldKeyedNodes) => {
-            currentMatchParentChain.forEach(itemKey => {
-                oldKeyedNodes[itemKey].collapsed = false;
-                oldKeyedNodes[itemKey].lastModifiedUnixMillis = Date.now();
-            })
-            return oldRoot;
-        })
-        if (currentMatchId != NO_MATCH) {
-            props.itemRefsDictionary[currentMatchId].focusThis();
-        } else {
-            props.itemRefsDictionary[props.lastFocusedItem].focusThis();
-        }
-    }
-    const scrollToCurrentItem = () => {
-        props.itemRefsDictionary[currentMatchId]?.scrollThisIntoView();
-    }
+    const [omniBarState, setOmniBarState] = React.useState<OmniBarState>(getDefaultOmnibarState());
+    const { omnibarKeyHandler, rootNode, extraAnnotations } = getSpecializedProps(omniBarState, setOmniBarState, props.transformedDataAndSetter, props.itemRefsDictionary);
     return <>
         <OmniBar
             omniBarState={omniBarState}
             setOmniBarState={setOmniBarState}
-            nMatches={nMatches}
-            focusOnCurrentItem={focusOnCurrentItem}
-            scrollToCurrentItem={scrollToCurrentItem}
-        ></OmniBar>
+            omnibarKeyHandler={omnibarKeyHandler}
+            lastFocusedItem={props.lastFocusedItem}
+        >
+            {extraAnnotations}
+        </OmniBar>
         <ModelContext.Provider value={rootNode}>
             {props.children}
         </ModelContext.Provider>
     </>
 }
 
-
-
 const OmniBar = (props: {
     omniBarState: OmniBarState,
     setOmniBarState: React.Dispatch<React.SetStateAction<OmniBarState>>,
-    focusOnCurrentItem: () => void,
-    scrollToCurrentItem: () => void,
-    nMatches: number
+    omnibarKeyHandler: (evt: React.KeyboardEvent) => void,
+    lastFocusedItem: string, // Todo: Elevate this property + the inputReference up
+    children: React.ReactElement
 }) => {
     const inputReference = React.useRef<HTMLInputElement>(null);
+    const lastFocusedItem_ = React.useRef<string>("");
     React.useEffect(() => {
         const listenForCtrlFP = (e: KeyboardEvent) => {
             if ((e.key == "f" || e.key == "p") && (e.ctrlKey || e.metaKey)) {
@@ -76,11 +49,13 @@ const OmniBar = (props: {
                     visibleChangeOccured = true;
                 }
                 if (e.key == "p") {
+                    // Save currently focused item
                     props.setOmniBarState((currentState) => {
                         if (!currentState.barContents.startsWith(">")) {
                             return {
                                 ...currentState,
-                                barContents: ">"
+                                barContents: ">",
+                                preOmnibarFocusItemId: lastFocusedItem_.current
                             }
                         } else {
                             return currentState;
@@ -95,36 +70,31 @@ const OmniBar = (props: {
         }
         window.addEventListener("keydown", listenForCtrlFP);
         return () => window.removeEventListener("keydown", listenForCtrlFP);
-    }, [inputReference.current]);
+    }, []);
 
-    const matchMessage = props.nMatches > 0 ? `${props.omniBarState.searchState.searchSelectionIdx + 1} / ${props.nMatches} matches` : "No matches"
+    React.useEffect(() => {
+        // TODO: Consider promoting this state to parent
+        if (props.omniBarState.barContents == "") {
+            lastFocusedItem_.current = props.lastFocusedItem;
+        }
+    }, [props.lastFocusedItem]);
 
     return <div className="search-bar">
         <input ref={inputReference}
             placeholder={"🔍 Search"}
             value={props.omniBarState.barContents}
             onChange={(evt) => props.setOmniBarState((oldState) => ({ ...oldState, barContents: evt.target.value }))}
-            onKeyDown={(evt) => {
-                if (evt.key == "ArrowUp") {
-                    props.setOmniBarState((oldState) => ({ ...oldState, searchState: { ...oldState.searchState, searchSelectionIdx: oldState.searchState.searchSelectionIdx - 1 } }))
-                    window.setTimeout(props.scrollToCurrentItem, 1);
-                } else if (evt.key == "ArrowDown") {
-                    props.setOmniBarState((oldState) => ({ ...oldState, searchState: { ...oldState.searchState, searchSelectionIdx: oldState.searchState.searchSelectionIdx + 1 } }))
-                    window.setTimeout(props.scrollToCurrentItem, 1);
-                } else if (evt.key == "Enter") {
-                    props.focusOnCurrentItem();
-                } else if (evt.key == "Escape") {
-                    props.focusOnCurrentItem();
-                } else {
-                    // Search query was changed
-                    window.setTimeout(props.scrollToCurrentItem, 1);
-                }
-            }}
-            style={{ flex: "1 0 auto", padding: "2px" }}></input>
-        {
-            props.omniBarState.barContents.length > 0 ?
-                <span>{matchMessage}</span> :
-                null
-        }
+            onKeyDown={props.omnibarKeyHandler}
+            style={{ flex: "1 0 auto", padding: "2px" }} />
+        {props.children}
     </div>
 }
+
+export const getDefaultOmnibarState = (): OmniBarState => {
+    return {
+        barContents: "",
+        selectionIdx: 0,
+    }
+}
+
+export default OmniBarWrapper;