+++ docs/changelog.md
@@ -3,6 +3,7 @@
 - Add CTRL+P for omnibar focus
 - Highlight currently selected item
 - Make Escape on omnibar return focus to last focused item
+- Refactor to make Omnibar jump work
 
 # Version 3.5.0: More useful search 
 - Fix minor focus bug